Fix scheduling issues and proactive Time message

+++ presentation/bot/utils/reactOnCommand.js
@@ -11,8 +11,9 @@ const MAIN_ITERATION_WATERFALL_DIALOG = 'MAIN_ITERATION_WATERFALL_DIALOG';
 module.exports.reactOnCommand = async function (command, turnContext, botState) {
     const iterationDialog = new IterationDialog(
         MAIN_ITERATION_WATERFALL_DIALOG,
-        async (result) =>
-            await onIterationDialogFinish(botState.dialogState, result)
+        botState,
+        async (result, state) =>
+            await onIterationDialogFinish(state.dialogState, result)
     );
 
     switch (command.name) {
@@ -39,20 +40,22 @@ async function onIterationDialogFinish(dialogState, result) {
     const iterationsRepo = Injection.getInstance('DAL.IterationRepository');
     const iterationsManager = Injection.getInstance('SkypeBot.NewIteration');
 
-    dialogState.lastDialog = null;
-
-    switch (result.action) {
-    case 'DELETE':
-        await iterationsRepo.remove(result.iteration.id);
-        break;
-    case 'EDIT':
-        await iterationsRepo.remove(result.previousId);
-        await iterationsManager.addIterations([result.iteration]);
-        break;
-    case 'ADD':
-        await iterationsManager.addIterations([result.iteration]);
-        break;
-    default:
-        break;
+    if (result) {
+        switch (result.action) {
+        case 'DELETE':
+            await iterationsRepo.remove(result.iteration.id);
+            break;
+        case 'EDIT':
+            await iterationsRepo.remove(result.previousId);
+            await iterationsManager.addIterations([result.iteration]);
+            break;
+        case 'ADD':
+            await iterationsManager.addIterations([result.iteration]);
+            break;
+        default:
+            break;
+        }
     }
+
+    dialogState.lastDialog = null;
 };
